+++ lib/platform/video_player_method_channel.dart
@@ -370,8 +370,8 @@ class MethodChannelVideoPlayer extends VideoPlayerPlatform {
             eventType: eventType,
             key: map["key"] as String?,
             progress: map["progress"] as double,
-            bytesDownloaded: map["bytesDownloaded"] as double?,
-            bytesTotal: map["bytesTotal"] as double?,
+            bytesDownloaded: map["bytesDownloaded"] as int?,
+            bytesTotal: map["bytesTotal"] as int?,
           );
         case PlatformDownloadEventType.finished:
           return PlatformDownloadEvent(
